4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -45,8 +45,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+  if (decodedToken.id === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'token invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
